Add --minimize and --maximize options
This requires expanding the opt_flags to 64 bit since we have
now used up all the 32 bits of the flags(!)

+++ log.c
@@ -41,7 +41,7 @@ bool	 abort_msg_emitted;
  */
 int print(
 	FILE *fp,
-	const int flag,
+	const uint64_t flag,
 	const char *const fmt, ...)
 {
 	va_list ap;
@@ -95,7 +95,7 @@ int print(
  *	print failure message with errno
  */
 void pr_failed(
-	const int flag,
+	const uint64_t flag,
 	const char *name,
 	const char *what,
 	const int err)
